update db model struct

+++ src/handler/menu/shop.go
@@ -53,7 +53,7 @@ func GetShop(c *gin.Context) {
 	}
 	shopName := c.Query("name")
 
-	data, err := logic.GetShop(int32(shopID), shopName)
+	data, err := logic.GetShop(shopID, shopName)
 	if err != nil {
 		c.AbortWithError(http.StatusBadRequest, err)
 		return
@@ -62,7 +62,7 @@ func GetShop(c *gin.Context) {
 	result := make([]resp.Shop, 0)
 	for _, v := range data {
 		result = append(result, resp.Shop{
-			ID:   v.GetID(),
+			ID:   int32(v.GetID()),
 			Name: v.GetName(),
 		})
 	}
@@ -88,7 +88,7 @@ func UpdateShop(c *gin.Context) {
 	}
 	shopName := c.PostForm("name")
 
-	success, err := logic.UpdateShop(int32(shopID), shopName)
+	success, err := logic.UpdateShop(shopID, shopName)
 	if err != nil {
 		c.AbortWithError(http.StatusBadRequest, err)
 		return
@@ -114,7 +114,7 @@ func DeleteShop(c *gin.Context) {
 		return
 	}
 
-	success, err := logic.DeleteShop(int32(shopID))
+	success, err := logic.DeleteShop(shopID)
 	if err != nil {
 		c.AbortWithError(http.StatusBadRequest, err)
 		return
